+++ cmd/deploy.go
@@ -83,6 +83,9 @@ func deployTemplate(cmd *cobra.Command, args []string) {
 		os.Exit(1)
 	}
 
+	// Capture deployment start timestamp before any AWS operations
+	deployment.DeploymentStart = time.Now()
+
 	deploymentLog := lib.NewDeploymentLog(awsConfig, deployment)
 
 	precheckOutput := runPrechecks(&deployment, &deploymentLog)
@@ -97,15 +100,20 @@ func deployTemplate(cmd *cobra.Command, args []string) {
 }
 
 // showDeploymentInfo shows what kind of deployment this (New/Update) and where it's happening
-func showDeploymentInfo(deployment lib.DeployInfo, awsConfig config.AWSConfig) {
+func showDeploymentInfo(deployment lib.DeployInfo, awsConfig config.AWSConfig, quiet bool) {
+	// Return early if quiet mode is enabled
+	if quiet {
+		return
+	}
+
 	bold := color.New(color.Bold).SprintFunc()
 	method := determineDeploymentMethod(deployment.IsNew, deployFlags.Dryrun)
 	account := formatAccountDisplay(awsConfig.AccountID, awsConfig.AccountAlias)
 
 	if deployment.IsNew {
-		fmt.Printf("%v new stack '%v' to region %v of account %v\n\n", method, bold(deployFlags.StackName), awsConfig.Region, account)
+		fmt.Fprintf(os.Stderr, "%v new stack '%v' to region %v of account %v\n\n", method, bold(deployFlags.StackName), awsConfig.Region, account)
 	} else {
-		fmt.Printf("%v stack '%v' in region %v of account %v\n\n", method, bold(deployFlags.StackName), awsConfig.Region, awsConfig.AccountID)
+		fmt.Fprintf(os.Stderr, "%v stack '%v' in region %v of account %v\n\n", method, bold(deployFlags.StackName), awsConfig.Region, awsConfig.AccountID)
 	}
 	printBasicStackInfo(deployment, true, awsConfig)
 }
@@ -228,7 +236,7 @@ func setDeployParameters(deployment *lib.DeployInfo) {
 }
 
 func createChangeset(deployment *lib.DeployInfo, awsConfig config.AWSConfig) *lib.ChangesetInfo {
-	if deployment.TemplateUrl != "" {
+	if deployment.TemplateUrl != "" && !deployFlags.Quiet {
 		text := fmt.Sprintf("Using template uploaded as %v", deployment.TemplateUrl)
 		printMessage(formatInfo(text))
 	}
@@ -251,8 +259,8 @@ func createChangeset(deployment *lib.DeployInfo, awsConfig config.AWSConfig) *li
 		}
 		// Otherwise, show the error and clean up
 		printMessage(formatError(string(texts.DeployChangesetMessageCreationFailed)))
-		fmt.Println(changeset.StatusReason)
-		fmt.Printf("\n%v %v\n", texts.DeployChangesetMessageConsole, changeset.GenerateChangesetUrl(awsConfig))
+		fmt.Fprintln(os.Stderr, changeset.StatusReason)
+		fmt.Fprintf(os.Stderr, "\n%v %v\n", texts.DeployChangesetMessageConsole, changeset.GenerateChangesetUrl(awsConfig))
 		var deleteChangesetConfirmation bool
 		if deployFlags.NonInteractive {
 			deleteChangesetConfirmation = true
@@ -293,7 +301,7 @@ func deleteChangeset(deployment lib.DeployInfo, awsConfig config.AWSConfig) {
 }
 
 func deleteStackIfNew(deployment lib.DeployInfo, awsConfig config.AWSConfig) {
-	fmt.Println(texts.DeployStackMessageNewStackDeleteInfo)
+	fmt.Fprintln(os.Stderr, texts.DeployStackMessageNewStackDeleteInfo)
 	var deleteStackConfirmation bool
 	if deployFlags.Dryrun || deployFlags.NonInteractive {
 		deleteStackConfirmation = true
@@ -314,39 +322,49 @@ func deleteStackIfNew(deployment lib.DeployInfo, awsConfig config.AWSConfig) {
 			}
 		}
 	} else {
-		fmt.Println("No problem. I have left the stack intact, please delete it manually once you're done.")
+		fmt.Fprintln(os.Stderr, "No problem. I have left the stack intact, please delete it manually once you're done.")
 	}
 }
 
 func deployChangeset(deployment lib.DeployInfo, awsConfig config.AWSConfig) {
-	if deployFlags.NonInteractive {
-		printMessage(formatInfo(string(texts.DeployChangesetMessageAutoDeploy)))
-	} else {
-		printMessage(formatSuccess(string(texts.DeployChangesetMessageWillDeploy)))
+	if !deployFlags.Quiet {
+		if deployFlags.NonInteractive {
+			printMessage(formatInfo(string(texts.DeployChangesetMessageAutoDeploy)))
+		} else {
+			printMessage(formatSuccess(string(texts.DeployChangesetMessageWillDeploy)))
+		}
 	}
 	err := deployment.Changeset.DeployChangeset(awsConfig.CloudformationClient())
 	if err != nil {
 		printMessage(formatError("Could not execute changeset! See details below"))
-		fmt.Println(err)
+		fmt.Fprintln(os.Stderr, err)
 	}
 	latest := deployment.Changeset.CreationTime
 	time.Sleep(3 * time.Second)
-	fmt.Println(formatBold("Showing the events for the deployment:"))
+	if !deployFlags.Quiet {
+		fmt.Fprintln(os.Stderr, formatBold("Showing the events for the deployment:"))
+	}
 	ongoing := true
 	for ongoing {
-		latest = showEvents(deployment, latest, awsConfig)
+		latest = showEvents(deployment, latest, awsConfig, deployFlags.Quiet)
 		time.Sleep(3 * time.Second)
 		ongoing = deployment.IsOngoing(awsConfig.CloudformationClient())
 	}
 	// One last time after the deployment finished in case of a timing mismatch
-	showEvents(deployment, latest, awsConfig)
+	showEvents(deployment, latest, awsConfig, deployFlags.Quiet)
 }
 
-func showEvents(deployment lib.DeployInfo, latest time.Time, awsConfig config.AWSConfig) time.Time {
+func showEvents(deployment lib.DeployInfo, latest time.Time, awsConfig config.AWSConfig, quiet bool) time.Time {
+	// Return early if quiet mode is enabled
+	if quiet {
+		return latest
+	}
+
 	events, err := deployment.GetEvents(awsConfig.CloudformationClient())
 	if err != nil {
 		printMessage(formatError("Something went wrong trying to get the events of the stack"))
-		fmt.Println(err)
+		fmt.Fprintln(os.Stderr, err)
+		return latest
 	}
 	sort.Sort(ReverseEvents(events))
 	for _, event := range events {
@@ -356,12 +374,12 @@ func showEvents(deployment lib.DeployInfo, latest time.Time, awsConfig config.AW
 			switch event.ResourceStatus {
 			case types.ResourceStatusCreateFailed, types.ResourceStatusImportFailed, types.ResourceStatusDeleteFailed, types.ResourceStatusUpdateFailed, types.ResourceStatusImportRollbackComplete, types.ResourceStatus(types.StackStatusRollbackComplete), types.ResourceStatus(types.StackStatusUpdateRollbackComplete):
 				// For streaming logs, just apply color without extra spacing
-				fmt.Println(output.StyleWarning(message))
+				fmt.Fprintln(os.Stderr, output.StyleWarning(message))
 			case types.ResourceStatusCreateComplete, types.ResourceStatusImportComplete, types.ResourceStatusUpdateComplete, types.ResourceStatusDeleteComplete:
 				// For streaming logs, just apply color without extra spacing
-				fmt.Println(output.StylePositive(message))
+				fmt.Fprintln(os.Stderr, output.StylePositive(message))
 			default:
-				fmt.Println(message)
+				fmt.Fprintln(os.Stderr, message)
 			}
 		}
 	}
@@ -372,7 +390,7 @@ func showFailedEvents(deployment lib.DeployInfo, awsConfig config.AWSConfig, pre
 	events, err := deployment.GetEvents(awsConfig.CloudformationClient())
 	if err != nil {
 		printMessage(formatError("Something went wrong trying to get the events of the stack"))
-		fmt.Println(err)
+		fmt.Fprintln(os.Stderr, err)
 		return nil
 	}
 	changesetkeys := []string{"CfnName", "Type", "Status", "Reason"}
@@ -410,9 +428,9 @@ func showFailedEvents(deployment lib.DeployInfo, awsConfig config.AWSConfig, pre
 		)
 
 		doc := builder.Build()
-		out := output.NewOutput(settings.GetOutputOptions()...)
+		out := createStderrOutput()
 		if err := out.Render(context.Background(), doc); err != nil {
-			fmt.Printf("ERROR: Failed to render failed events: %v\n", err)
+			fmt.Fprintf(os.Stderr, "ERROR: Failed to render failed events: %v\n", err)
 		}
 	} else if prefixMessage != "" {
 		// If no failed events but we have a prefix message, still show it